ndslice::view

Struct Region

pub struct Region { /* private fields */ }

Region describes a region of a possibly-larger space of ranks, organized into a hyperrect.

Internally, region consist of a set of labels and a Slice, as it allows for a compact but useful representation of the ranks. However, this representation may change in the future.

Implementations

impl Region

pub fn new(labels: Vec<String>, slice: Slice) -> Self

Constructor to build arbitrary regions (incl. non-contiguous / offset).

pub fn labels(&self) -> &[String]

The labels of the dimensions of this region.

pub fn slice(&self) -> &Slice

The slice representing this region. Note: this representation may change.

pub fn into_inner(self) -> (Vec<String>, Slice)

Convert this region into its constituent labels and slice.

pub fn extent(&self) -> Extent

Returns the extent of the region.

pub fn is_subset(&self, other: &Region) -> bool

Returns true if this region is a subset of other, i.e., if other contains at least all of the ranks in this region.

pub fn remap(&self, target: &Region) -> Option<impl Iterator<Item = usize> + '_>

Remap the target to ranks in this region. The returned iterator iterates over each rank in target, providing the corresponding rank in self. This is useful when mapping between different subspaces.

let ext = extent!(replica = 8, gpu = 4);
let replica1 = ext.range("replica", 1).unwrap();
assert_eq!(replica1.extent(), extent!(replica = 1, gpu = 4));
let replica1_gpu12 = replica1.range("gpu", 1..3).unwrap();
assert_eq!(replica1_gpu12.extent(), extent!(replica = 1, gpu = 2));
// The first rank in `replica1_gpu12` is the second rank in `replica1`.
assert_eq!(
    replica1.remap(&replica1_gpu12).unwrap().collect::<Vec<_>>(),
    vec![1, 2],
);

pub fn num_ranks(&self) -> usize

Returns the total number of ranks in the region.

pub fn base_rank_of_point(&self, p: Point) -> Result<usize, RegionError>

Convert a rank in this region’s extent into its corresponding rank in the base space defined by the region’s Slice.

pub fn point_of_base_rank(&self, rank: usize) -> Result<Point, RegionError>

Convert a rank in the base space into the corresponding Point in this region’s extent (if the base rank lies within the region’s Slice).