ndslice::view

Trait View

pub trait View: Sized {
    type Item;
    type View: View;

    // Required methods
    fn region(&self) -> Region;
    fn get(&self, rank: usize) -> Option<Self::Item>;
    fn subset(&self, region: Region) -> Result<Self::View, ViewError>;
}

A View is a collection of items in a space indexed by a Region.

Required Associated Types

type Item

The type of item in this view.

type View: View

The type of sub-view produced by manipulating (e.g., slicing) this view.

Required Methods

fn region(&self) -> Region

The ranks contained in this view.

fn get(&self, rank: usize) -> Option<Self::Item>

Retrieve the item corresponding to the given rank in the Region of this view. An implementation MUST return a value for all ranks defined in this view.

fn subset(&self, region: Region) -> Result<Self::View, ViewError>

Subsets this view with the provided ranks. This is mainly used by combinators on Views themselves. The set of ranks passed in must be a subset of the ranks of the base view.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.